Added new version of 'library.title' to return the title of the media excluding year and identifiers.
