+++ couchpotato/core/media/_base/searcher/main.py
@@ -171,13 +171,7 @@ class Searcher(SearcherBase):
         return False
 
     def correctWords(self, rel_name, media):
-        media_title = fireEvent(
-            'library.title', media['library'],
-            condense = False,
-            include_year = False,
-            include_identifier = False,
-            single = True
-        )
+        media_title = fireEvent('library.title', media['library'], single = True)
         media_words = re.split('\W+', simplifyString(media_title))
 
         rel_name = simplifyString(rel_name)
